+++ Tools/JobDeployK8s.tests.ps1
@@ -78,4 +78,11 @@ Describe "Deploys Minimal API" {
         $podStatus.Status | Should -Be 'Completed'
         kubectl delete job 'test-job' --ignore-not-found
     } -Tag 'Happy','k19'
+
+    It "has a bad tag on the container" {
+        $deploy = Deploy-MinimalJobK8s -SkipInit -ImageTag "latest " -TimeoutSecs 5
+        $deploy.Status | Should -Be 'ConfigError'
+        $deploy.LastBadEvents.Count | Should -BeGreaterThan 0
+        'Error creating: Pod "..." is invalid: spec.containers[0].image: Invalid value: "init-app:latest ": must not have leading or trailing whitespace' | Should -BeIn $deploy.LastBadEvents
+    } -Tag 'Sad', 'Config', 'k20'
 }